Immediately record what's in sitemap, improve scan
Previously, we'd only been recording that something was in the sitemap
if we'd also decided to scan it in Google Search for this run - as we
can only scan a very few things in Google Search with each run, that
wasn't great at getting us a picture of what the newest arrivals in the
sitemap were.

We want to prefer scanning new arrivals as they're most likely to be
missing, and it's probably not worth wasting Google Search API quota on
old entries when there are 100s of them and most of them are probably
fine - climbing that hill on first startup is especially arduous!

Even over new arrivals we also want to priortise ones that are known
missing from Google results- and haven't been scanned for the longest
time.