Avoid reporting on weird long-lived NYT pages
The NYT have several links in their 'news' sitemap (https://www.nytimes.com/sitemaps/new/news-1.xml.gz)
which are long-lived, and often weird in different ways (eg, this page
redirects, but does it with Javascript rather than HTTP response code:
https://www.nytimes.com/interactive/2022/world/ukraine-live.html
). The weirdness is a pain for the observatory to work with, but
thankfully we're not really interested in _old_ content - we just want
to know how long it takes urgent *news* to turn up in Google Search -
so this change filters out everything that was first seen in the
sitemap more than 4 hours ago, so we only scan stuff for the first 4
hours of it's life.

This should also lower API consumption, which is nice.