Allow scanning missing content a little more often
<- Multi-line description, try to limit each line to max 72 chars -->|
There's a hard cap on the number of missing content that we scan each
time, but we do prioritise scanning the content that's gone the
longest without a scan. The upshot is that we'll only do really
*frequent* scanning if only a few articles are missing at a time -
if there are instead a *lot* to scan (because Google's gone wrong
again, and lots are missing from Google Search results) we'll not
expend vast amounts of API quota.