Reduce API consumption cost for long-missed content
Currently we're scanning 3 news sites (BBC, NYT, Daily Mail), and this
is consuming ~7500 Google Search API calls per day - so maybe ~2500
calls per day per site, or ~$4500 per site per year in API costs. That's
a fair bit of cash, and there's already interest to add at least 4 other
sites (Washpo, Telegraph, Sun, and ABC for Dave Earley in Australia).
Reducing costs would be good.

For content that's missing for a long period of time - say 1 hour - we
can use up to 30 API calls just scanning the content every 2 minutes.
However, as the length of time the content is missing increases, the
chances that it will suddenly appear become lower, as does our interest
in being precise about exactly how long it was missing for. Once content
has been missing for an hour, a reasonable time between checks to see
if it's still missing might be 10 minutes or more. This reasonable time
between checks increases the longer the content has been missing for.
With this update, content missing for 1 hour will take fewer than 12
API calls (reducing cost by 60%), and missing for 2 hours fewer than
15 in total (reducing cost by 75%).

If a piece of content has never been seen before in the sitemap, we
still want to scan it within a minute.

For the fastest possible check times - which won't quite happen in
practice, as we only run once per minute - the elapsed time at each
check is now :

1M, 3M12S, 5M50S, 9M, 12M48S, 17M22S, 22M50S, 29M24S, 37M17S, 46M45S, 58M6S, 1H11M43S, 1H28M4S, 1H47M41S, 2H11M13S, 2H39M28S, 3H13M22S, 3H54M2S