Add redirect-resolving code
The BBC have share location urls in their sitemap which redirect, eg:

https://www.bbc.co.uk/news/uk-politics-63534039 - in sitemap
https://www.bbc.co.uk/news/av/uk-politics-63534039 - ultimate url

...it's only the ultimate url (with '/av', in this case) that can be
found in Google Search results.

Updated process
---------------

Resolution of new urls only occurs when the url is first found in the
sitemap (we don't want to have to repeatedly resolve the same urls
over and over again) - this redirected url is stored in the new
`finalUriAfterRedirects` column.

Urls are never searched for in Google Search when we first see them in
the sitemap - we wait until we have had a new lambda run, and the url
has been loaded as an existing `AvailabilityRecord` in dynamodb.
It's at that point that we search for the redirected url, which we
recorded earlier.

Which url to use when?
----------------------

* original url : needed as the primary key, because the observatory
  lambda uses it to lookup every entry from the sitemap (we don't want
  to have to resolve every url from the sitemap every time we load the sitemap)
* redirected url : when we search google for the content, we need to be using
  the redirected url when we're checking for matches.

Added logging of bad HTTP-response status codes

As suggested by Sam!