Fix logging of what urls are stored
While trying to debug why https://github.com/guardian/ophan/pull/5560
hadn't fixed the big gap in data seen on 25th September, I tried to
analyse the logging of how many new urls were being stored to dynamodb,
but couldn't, because that particualr log line mysteriously wasn't
visible anywhere in the ELK:

https://logs.gutools.co.uk/s/ophan/goto/4e86eda0-5dea-11ee-87e8-6b5e97480045

It turns out that the logging was getting rejected at the stage of
ingestion by the ELK - it _was_ being successfully written out as text
to Cloudwatch Logs:

https://eu-west-1.console.aws.amazon.com/cloudwatch/home?region=eu-west-1#logsV2:log-groups/log-group/%2Faws%2Flambda%2Fophan-PROD-google-search-i-scheduledLambda8A84450D-YaBDtsuYvFLP/log-events/2023%2F09%2F28%2F[%24LATEST]68633a68afe34703a8f60893d32b3482?start=PT3H

...but when that was being picked up by https://github.com/guardian/cloudwatch-logs-management
and sent to the ELK, the log context data was being rejected as bad.

In the ELK, the dot-nesting we make for context fields actually defines
a real Elasticsearch document structure, and in that structure, a
particular path can be either a single value, or a containing object,
but not both. The ELK works out which from the values we give it-
but sometimes those implications contradict each other:

"site" -> sitemapDownload.site.url ...means `site` is a string value
"sitemap.uris.old" -> alreadyKnownUris.size ...means `site` is an _object_ that contains other values

This is a 'mapping conflict', and the whole log line gets dropped:

https://discuss.elastic.co/t/tried-to-parse-field-as-object-but-found-a-concrete-value/229134

This is a similar issue to https://github.com/guardian/ophan/pull/3254
seen in the past. It would be cool if the logging code within the lambda
itself could identify issues like this and warn about them!